Hoist 'app_engine_id'/'compute_engine_id' into 'gcloud._helpers'.
